A function that prints an integer

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -1,4 +1,5 @@
 #include "main.h"
+#include <stdio.h>
 
 /**
  * print_number- prints an integer
@@ -17,5 +18,5 @@ void print_number(int n)
 	{
 		print_number(num / 10);
 	}
-	_putchar((num % 10) + '0');
+	_putchar((unsigned int)num % 10 + '0');
 }
